Refactor to remove dependency on github.com/moby/moby/pkg/progress
