+++ cli/command/container/run_test.go
@@ -12,10 +12,10 @@ import (
 	"github.com/creack/pty"
 	"github.com/docker/cli/cli"
 	"github.com/docker/cli/cli/streams"
+	"github.com/docker/cli/internal/progress"
 	"github.com/docker/cli/internal/streamformatter"
 	"github.com/docker/cli/internal/test"
 	"github.com/docker/cli/internal/test/notary"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/network"
