+++ cli/command/image/build.go
@@ -19,10 +19,10 @@ import (
 	"github.com/docker/cli/cli/command/image/build"
 	"github.com/docker/cli/cli/streams"
 	"github.com/docker/cli/internal/jsonstream"
+	"github.com/docker/cli/internal/progress"
 	"github.com/docker/cli/internal/streamformatter"
 	"github.com/docker/cli/opts"
 	"github.com/moby/go-archive"
-	"github.com/moby/moby/api/pkg/progress"
 	buildtypes "github.com/moby/moby/api/types/build"
 	"github.com/moby/moby/api/types/container"
 	registrytypes "github.com/moby/moby/api/types/registry"
