+++ cli/command/service/progress/progress.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/docker/cli/cli/command/formatter"
+	"github.com/docker/cli/internal/progress"
 	"github.com/docker/cli/internal/streamformatter"
-	"github.com/moby/moby/api/pkg/progress"
 	"github.com/moby/moby/api/types/swarm"
 	"github.com/moby/moby/client"
 )
