+++ cli/command/service/progress/progress_test.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/moby/moby/api/pkg/progress"
+	"github.com/docker/cli/internal/progress"
 	"github.com/moby/moby/api/types/swarm"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
